Adds more items to winform

+++ Bookstore.Domain/Customer.cs
@@ -3,6 +3,6 @@
     public class Customer : Person
     {
         public required int Cust_Id { get; set; }
-        public required int Phone_No { get; set; }
+        public required long Phone_No { get; set; }
     }
 }
